refactor: OneMap helper to TS
