+++ src/onemap/onemap.ts
@@ -1,0 +1,39 @@
+import axios from 'axios';
+import querystring from 'querystring';
+
+const BASE_URL = 'https://developers.onemap.sg';
+
+export interface Response {
+  found: number;
+  totalNumPages: number;
+  pageNum: number;
+  results: Result[];
+}
+
+export interface Result {
+  SEARCHVAL: string;
+  X: string;
+  Y: string;
+  LATITUDE: string;
+  LONGITUDE: string;
+  LONGTITUDE: string;
+}
+
+/**
+ * Search using Onemap.sg
+ */
+export default async function search(term: string): Promise<Response> {
+  const query = {
+    searchVal: term,
+    returnGeom: 'Y',
+    getAddrDetails: 'N',
+    pageNum: '1',
+  };
+  const response = await axios.get(
+    `${BASE_URL}/commonapi/search?${querystring.stringify(query)}`,
+    {
+      responseType: 'json',
+    }
+  );
+  return response.data;
+}
